refactor(domain): store monetary amounts as int64 cents

+++ billsplit/internal/domain/splits_test.go
@@ -9,26 +9,26 @@ import (
 )
 
 func TestValidateSplits_Valid(t *testing.T) {
-	err := domain.ValidateSplits(100, map[string]float64{"alice": 30, "bob": 70}, []string{"alice", "bob"})
+	err := domain.ValidateSplits(100, map[string]int64{"alice": 30, "bob": 70}, []string{"alice", "bob"})
 	require.NoError(t, err)
 }
 
 func TestValidateSplits_DoesNotSum(t *testing.T) {
-	err := domain.ValidateSplits(100, map[string]float64{"alice": 30, "bob": 60}, []string{"alice", "bob"})
+	err := domain.ValidateSplits(100, map[string]int64{"alice": 30, "bob": 60}, []string{"alice", "bob"})
 	require.Error(t, err, "expected error: splits don't sum to total")
 }
 
 func TestValidateSplits_UnknownMember(t *testing.T) {
-	err := domain.ValidateSplits(100, map[string]float64{"alice": 50, "carol": 50}, []string{"alice", "bob"})
+	err := domain.ValidateSplits(100, map[string]int64{"alice": 50, "carol": 50}, []string{"alice", "bob"})
 	require.Error(t, err, "expected error: unknown member in splits")
 }
 
 func TestValidateSplits_NegativeAmount(t *testing.T) {
-	err := domain.ValidateSplits(100, map[string]float64{"alice": 110, "bob": -10}, []string{"alice", "bob"})
+	err := domain.ValidateSplits(100, map[string]int64{"alice": 110, "bob": -10}, []string{"alice", "bob"})
 	require.Error(t, err, "expected error: negative split")
 }
 
 func TestValidateSplits_Empty(t *testing.T) {
-	err := domain.ValidateSplits(100, map[string]float64{}, []string{"alice", "bob"})
+	err := domain.ValidateSplits(100, map[string]int64{}, []string{"alice", "bob"})
 	require.Error(t, err, "expected error: empty splits")
 }
